Изменить иерархию и улучшить тесты
