+++ src/main/java/ru/practicum/shareit/request/service/ItemRequestServicelmpl.java
@@ -9,8 +9,8 @@ import ru.practicum.shareit.request.repository.RequestRepository;
 import ru.practicum.shareit.user.model.User;
 import  ru.practicum.shareit.request.dto.ItemRequestMapper;
 import ru.practicum.shareit.user.repository.UserRepository;
-import ru.practicum.shareit.validator.RequestValidator;
-import ru.practicum.shareit.validator.UserValidator;
+import ru.practicum.shareit.validator.RequestValidatorService;
+import ru.practicum.shareit.validator.UserValidatorService;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
@@ -24,8 +24,8 @@ import static ru.practicum.shareit.request.dto.ItemRequestMapper.*;
 public class ItemRequestServicelmpl implements ItemRequestService {
     private final RequestRepository requestRepository;
     private final UserRepository userRepository;
-    private final UserValidator userValidator;
-    private final RequestValidator requestValidator;
+    private final UserValidatorService userValidator;
+    private final RequestValidatorService requestValidator;
 
     @Override
     public ItemRequestDto addNewRequest(ItemRequestDto requestDto, Long userId) {
@@ -38,7 +38,7 @@ public class ItemRequestServicelmpl implements ItemRequestService {
 
     @Override
     public Collection<ItemRequestDto> getAllUserRequestsWithResponses(Long userId) {
-        userValidator.checkingUserIdAndNotReturns(userId);
+        userValidator.checkingUserIdAndNotReturn(userId);
         return requestRepository.findAllByRequester_Id(userId).stream()
                 .map(ItemRequestMapper::toItemRequestDto)
                 .collect(Collectors.toList());
@@ -46,7 +46,7 @@ public class ItemRequestServicelmpl implements ItemRequestService {
 
     @Override
     public Collection<ItemRequestDto> getAllRequestsToResponse(Long userId, Pageable page) {
-        userValidator.checkingUserIdAndNotReturns(userId);
+        userValidator.checkingUserIdAndNotReturn(userId);
         return requestRepository.findAllByAllOtherUsers(userId, page).stream()
                 .map(ItemRequestMapper::toItemRequestDto)
                 .collect(Collectors.toList());
@@ -54,8 +54,8 @@ public class ItemRequestServicelmpl implements ItemRequestService {
 
     @Override
     public ItemRequestDto getRequestById(Long userId, Long requestId) {
-        userValidator.checkingUserIdAndNotReturns(userId);
-        ItemRequest request = requestValidator.validateItemRequestIdAndReturns(requestId);
+        userValidator.checkingUserIdAndNotReturn(userId);
+        ItemRequest request = requestValidator.validateItemRequestIdAndReturn(requestId);
         return toItemRequestDto(request);
     }
 }
